Fixed SmsService now is returning string token and adding new SmsAuthantication object Changes on the db entities which had int for id now they have long for id Changes on the account controller now the register is working with binding model which has phone number and device and added smsverification action which is calling the sms service

+++ MRSServers/MRSMobile.Data.Models/MrsMobileDevice.cs
@@ -5,7 +5,7 @@ using System.ComponentModel.DataAnnotations;
 
 namespace MRSMobile.Data.Models
 {
-    public class MrsMobileDevice : BaseDeletableModel<int>
+    public class MrsMobileDevice : BaseDeletableModel<long>
     {
         public MrsMobileDevice()
         {
